Code cleanup for min/max functions.  Committed on behalf of lebedev.ri (thanks!).

+++ rtengine/improcfun.h
@@ -165,7 +165,7 @@ class ImProcFunctions {
 		static void rgb2hsv (float r, float g, float b, float &h, float &s, float &v);
 		static void hsv2rgb (float h, float s, float v, float &r, float &g, float &b);
 		void xyz2srgb (float x, float y, float z, float &r, float &g, float &b);
-		void xyz2rgb (float x, float y, float z, float &r, float &g, float &b, float rgb_xyz[3][3]);
+		void xyz2rgb (float x, float y, float z, float &r, float &g, float &b, double rgb_xyz[3][3]);
 		void Lab2XYZ(float L, float a, float b, float &x, float &y, float &z);
 		void XYZ2Lab(float X, float Y, float Z, float &L, float &a, float &b);
 		void Lab2Yuv(float L, float a, float b, float &Y, float &u, float &v);
